improved virtualizer scroll and column control
